feat(service): implement DTOs for user and password management
