+++ internal/service/dto/password.go
@@ -1,21 +1,9 @@
-package model
+package dto
 
 import (
+	"key-haven-back/internal/domain/user"
 	"time"
 )
-
-// Password representa as credenciais salvas de um usuário no sistema.
-type Password struct {
-	ID          string    `json:"id" bson:"_id,omitempty"`
-	UserID      string    `json:"user_id" bson:"user_id"`
-	Password    string    `json:"-" bson:"password"`
-	Login       string    `json:"login" bson:"login"`
-	Title       string    `json:"title" bson:"title"`
-	Description string    `json:"description" bson:"description"`
-	URL         string    `json:"url" bson:"url"`
-	CreatedAt   time.Time `json:"created_at" bson:"created_at"`
-	UpdatedAt   time.Time `json:"updated_at" bson:"updated_at"`
-}
 
 type CreatePasswordRequest struct {
 	UserID      string `json:"user_id" validate:"required"`
@@ -36,9 +24,9 @@ type CreatePasswordResponse struct {
 
 // NewPassword cria uma nova instância de Password a partir de um CreatePasswordRequest.
 // Gera um UUID para o ID e define os timestamps de criação e atualização.
-func NewPassword(req *CreatePasswordRequest) *Password {
+func NewPassword(req *CreatePasswordRequest) *user.Password {
 	now := time.Now()
-	return &Password{
+	return &user.Password{
 		UserID:      req.UserID,
 		Password:    req.Password,
 		Login:       req.Login,
