+++ internal/service/dto/user.go
@@ -1,0 +1,36 @@
+package dto
+
+import (
+	"key-haven-back/internal/domain/user"
+	"time"
+
+	"github.com/google/uuid"
+)
+
+type CreateUserRequest struct {
+	Email    string `json:"email" validate:"required,email"`
+	Password string `json:"password" validate:"required,min=12"`
+	Name     string `json:"name" validate:"required"`
+}
+
+type LoginRequest struct {
+	Email    string `json:"email" validate:"required,email"`
+	Password string `json:"password" validate:"required"`
+}
+
+type LoginResponse struct {
+	Token string    `json:"token"`
+	User  user.User `json:"user"`
+}
+
+func NewUser(req *CreateUserRequest) *user.User {
+	now := time.Now()
+	return &user.User{
+		ID:          uuid.New().String(),
+		Email:       req.Email,
+		Name:        req.Name,
+		CreatedAt:   now,
+		UpdatedAt:   now,
+		LastLoginAt: now,
+	}
+}
