+++ internal/service/password.go
@@ -2,12 +2,13 @@ package service
 
 import (
 	"context"
-	"key-haven-back/internal/model"
+	"key-haven-back/internal/domain/user"
 	"key-haven-back/internal/repository"
+	"key-haven-back/internal/service/dto"
 )
 
 type PasswordService interface {
-	Register(ctx context.Context, request *model.CreatePasswordRequest) (*model.Password, error)
+	Register(ctx context.Context, request *dto.CreatePasswordRequest) (*user.Password, error)
 }
 
 type passwordService struct {
@@ -20,6 +21,6 @@ func NewPasswordService(repo repository.PasswordRepository) PasswordService {
 	}
 }
 
-func (s *passwordService) Register(ctx context.Context, request *model.CreatePasswordRequest) (*model.Password, error) {
+func (s *passwordService) Register(ctx context.Context, request *dto.CreatePasswordRequest) (*user.Password, error) {
 	return nil, nil
 }
